pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "golem"
version = "0.1.0"
description = "MDAE pretraining for Graph Transformers on molecular descriptors"
requires-python = ">=3.10"
dependencies = [
    "torch>=1.13.0",
    "torch_geometric",
    "mordredcommunity",
    "rdkit",
    "dimorphite_dl",
    "numpy",
    "pandas",
    "pyyaml",
    "tqdm",
    "click",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov", "jupyter"]

[project.scripts]
golem = "golem.cli:main"

[tool.setuptools.packages.find]
include = ["golem*"]
